WIP - cleaned up the code base
