+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsmanageadjudicationsapi/config/WebClientConfig.kt
@@ -6,13 +6,17 @@ import org.springframework.beans.factory.annotation.Value
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
 import org.springframework.security.oauth2.client.OAuth2AuthorizedClientManager
+import org.springframework.security.oauth2.client.OAuth2AuthorizedClientProviderBuilder
+import org.springframework.security.oauth2.client.registration.ClientRegistrationRepository
+import org.springframework.security.oauth2.client.web.DefaultOAuth2AuthorizedClientManager
+import org.springframework.security.oauth2.client.web.OAuth2AuthorizedClientRepository
 import org.springframework.web.reactive.function.client.WebClient
 import uk.gov.justice.hmpps.kotlin.auth.authorisedWebClient
 import java.time.Duration
 
 @Configuration
 class WebClientConfig(
-  @Value("\${oauth.endpoint.url}") val authBaseUri: String,
+  @Value("\${hmpps.auth.url}") val authBaseUri: String,
   @Value("\${prison.nomis.location.api.endpoint.url}") val prisonNomisBaseUri: String,
   @Value("\${prison.location.api.endpoint.url}") val prisonLocationDetailBaseUri: String,
   @Value("\${prison.prisoner-search.api.endpoint.url}") val prisonerSearchBaseUri: String,
@@ -44,4 +48,22 @@ class WebClientConfig(
   fun prisonerSearchWebClient(authorizedClientManager: OAuth2AuthorizedClientManager, builder: WebClient.Builder) = builder
     .authorisedWebClient(authorizedClientManager, "prisoner-search-api", prisonerSearchBaseUri, apiTimeout)
     .also { log.info("WEB CLIENT CONFIG: creating prisoner search api web client") }
+
+  @Bean
+  fun authorizedClientManager(
+    clientRegistrationRepository: ClientRegistrationRepository,
+    authorizedClientRepository: OAuth2AuthorizedClientRepository
+  ): OAuth2AuthorizedClientManager {
+    val authorizedClientProvider = OAuth2AuthorizedClientProviderBuilder.builder()
+      .clientCredentials()
+      .build()
+
+    val authorizedClientManager = DefaultOAuth2AuthorizedClientManager(
+      clientRegistrationRepository,
+      authorizedClientRepository
+    )
+    authorizedClientManager.setAuthorizedClientProvider(authorizedClientProvider)
+
+    return authorizedClientManager
+  }
 }
